Upstream version 2.0.1

+++ libqpdf/QPDF_encryption.cc
@@ -10,6 +10,8 @@
 #include <qpdf/RC4.hh>
 #include <qpdf/MD5.hh>
 
+#include <string.h>
+
 static char const padding_string[] = {
     0x28, 0xbf, 0x4e, 0x5e, 0x4e, 0x75, 0x8a, 0x41,
     0x64, 0x00, 0x4e, 0x56, 0xff, 0xfa, 0x01, 0x08,
@@ -24,7 +26,7 @@ static unsigned int const key_bytes = 32;
 void
 pad_or_truncate_password(std::string const& password, char k1[key_bytes])
 {
-    int password_bytes = std::min(key_bytes, password.length());
+    int password_bytes = std::min((size_t) key_bytes, password.length());
     int pad_bytes = key_bytes - password_bytes;
     memcpy(k1, password.c_str(), password_bytes);
     memcpy(k1 + password_bytes, padding_string, pad_bytes);
@@ -425,7 +427,7 @@ QPDF::decryptStream(Pipeline*& pipeline, int objid, int generation,
 void
 QPDF::compute_encryption_O_U(
     char const* user_password, char const* owner_password,
-    int V, int R, int key_len, unsigned long P,
+    int V, int R, int key_len, int P,
     std::string const& id1, std::string& O, std::string& U)
 {
     EncryptionData data(V, R, key_len, P, "", "", id1);
